Move run to its own module
