+++ leverage/leverage.py
@@ -2,12 +2,10 @@
     Binbash Leverage Command-line Task Runner
 """
 
-import re
 import sys
 from pathlib import Path
 from importlib import util
 from inspect import getmembers
-from operator import attrgetter
 
 from leverage import __version__
 
@@ -15,82 +13,17 @@ from .task import Task
 from .task import NotATaskError
 from .task import MissingParensInDecoratorError
 from .path import NotARepositoryError
-from .path import NoBuildScriptFoundError
 from .path import get_build_script_path
-from .logger import get_logger
-from .logger import _attach_build_handler
-from ._parsing import _parse_args
-from ._parsing import InvalidArgumentOrderError
-from ._parsing import DuplicateKeywordArgumentError
-
-
-_TASK_PATTERN = re.compile(r"^(?P<name>[^\[\],\s]+)(\[(?P<arguments>[^\]]*)\])?$")
-_CREDIT_LINE = f"Powered by Leverage {__version__} - A Lightweight Python Build Tool based on Pynt."
-
-
-_logger = get_logger(name="leverage", level="INFO")
-_build_logger = get_logger(name="build", level="INFO")
-
-
-class MalformedTaskArgumentError(RuntimeError):
-    pass
-
-
-class TaskNotFoundError(RuntimeError):
-    pass
-
-
-def build(args):
-    """ Parse received arguments and execute the required action.
-
-    Args:
-        args (list): Arguments as received from command line.
-    """
-    parser = _get_argument_parser()
-    args = parser.parse_args(args=args)
-
-    # -v | --version
-    if args.version:
-        _print_version()
 
     # Load build file as a module
     try:
-        build_script = Path(get_build_script_path(filename=args.file))
-        module = _load_build_script(build_script=build_script)
 
     except (NotARepositoryError,
-            NoBuildScriptFoundError,
             NotATaskError,
             MissingParensInDecoratorError) as exc:
-        _terminate(error_message=str(exc))
-
-    # Attach a special filter as to add the name of the build script to every log record
-    _attach_build_handler(logger=_build_logger, build_script_name=module["name"])
-
-    # -h | --help
-    if args.help:
-        parser.print_help()
-    # -l
-    elif args.list_tasks:
-        _print_tasks(module)
-    # no args - run default task or print available tasks
-    elif not args.tasks:
-        default_task = module["__DEFAULT__"]
-        if default_task is not None:
-            prepared_default_task = [(default_task, [], {})]
-            _run_tasks(prepared_default_task)
-        else:
-            _print_tasks(module)
-    # tasks - run the given tasks
     else:
-        try:
-            tasks_to_run = _prepare_tasks_to_run(module, args.tasks)
 
-        except (TaskNotFoundError,
-                MalformedTaskArgumentError) as exc:
-            _terminate(error_message=str(exc))
 
-        _run_tasks(tasks=tasks_to_run)
 
 def _load_build_script(build_script):
     """ Load build script as module and return the useful bits.
@@ -131,98 +64,6 @@ def _get_tasks(module):
     # module (as the created task an as the `__DEFAULT__` variable value),
     # hence, the set, used to avoid repetition
     return list({task for _, task in getmembers(module, Task.is_task)})
-
-def _prepare_tasks_to_run(module, input_tasks):
-    """ Validate input tasks and arguments and pair them with the corresponding module's task.
-
-    Args:
-        module (dict): Dict containing the tasks from the build script.
-        input_tasks (list): Strings containing the tasks to invoke and their arguments as received
-            from user input.
-
-    Raises:
-        MalformedTaskArgumentError: When the string representing the invocation of a task does not conform
-            to the required pattern.
-        TaskNotFoundError: When the specified task is not found in the ones defined in the build script.
-        AmbiguousTaskNameError: When a given task name matches with multiple defined tasks.
-
-    Returns:
-        list(tuple): List of tasks paired with their corresponding args and kwargs as provided by the user.
-    """
-    tasks = []
-    for input_task in input_tasks:
-        match = _TASK_PATTERN.match(input_task)
-        if not match:
-            raise MalformedTaskArgumentError(f"Malformed task argument in `{input_task}`.")
-
-        name = match.group("name")
-        arguments = match.group("arguments")
-
-        try:
-            args, kwargs = _parse_args(arguments=arguments)
-
-        except (InvalidArgumentOrderError,
-                DuplicateKeywordArgumentError) as exc:
-            _terminate(error_message=str(exc).format(task=name))
-
-        task = [task for task in module["tasks"] if task.name == name]
-
-        if not task:
-            raise TaskNotFoundError(f"Unrecognized task `{name}`.")
-
-        tasks.append((task[0], args, kwargs))
-
-    return tasks
-
-def _run_tasks(tasks):
-    """ Run the tasks provided.
-
-    Args:
-        tasks (list(tuple)): List of 3-tuples containing a task and it's positional and keyword arguments to run.
-    """
-    completed_tasks = set()
-    for task, args, kwargs in tasks:
-        # Remove current task from dependencies set to force it to run, as it was invoked by the user explicitly
-        completed_tasks.discard(task)
-        _run(task, completed_tasks, *args, **kwargs)
-
-def _run(task, completed_tasks, *args, **kwargs):
-    """ Run the given task and all it's required dependencies, keeping track of all the already
-    completed tasks as not to repeat them.
-
-    Args:
-        task (list): Tasks to run.
-        completed_tasks (set): Tasks that have already ran.
-
-    Returns:
-        set: Updated set of already executed tasks.
-    """
-    # Satisfy dependencies recursively.
-    for dependency in task.dependencies:
-        _completed_tasks = _run(dependency, completed_tasks)
-        completed_tasks.update(_completed_tasks)
-
-    if task not in completed_tasks:
-
-        if task.is_ignored:
-            _build_logger.info(f"Ignoring task `{task.name}`")
-
-        else:
-            _build_logger.info(f"Starting task `{task.name}`")
-
-            try:
-                task(*args,**kwargs)
-            except:
-                _build_logger.critical(f"Error in task `{task.name}`")
-                _build_logger.critical("Aborting build")
-                raise
-
-            _build_logger.info(f"Completed task `{task.name}`")
-
-        completed_tasks.add(task)
-
-    return completed_tasks
-
 def _print_version():
     """ Print leverage version and quit. """
     _logger.info(f"leverage {__version__}")
